move dedup logic into jq parser
