+++ jq_service_parser.go
@@ -104,3 +104,17 @@ func (p *JQServiceParser) Run(json string) (*ServiceRegistration, error) {
 		Repositories: Repositories,
 	}, nil
 }
+
+func Deduplicated[T any](objects []T, keyFunc func(object T) string) []T {
+	out := make([]T, 0)
+	set := make(map[string]struct{})
+	for _, obj := range objects {
+		key := keyFunc(obj)
+		if _, ok := set[key]; ok {
+			continue
+		}
+		set[key] = struct{}{}
+		out = append(out, obj)
+	}
+	return out
+}
